// FIL/FIL.Web/ClientApp/models/UpdateTransactionResponseViewModel.ts
// This file was generated from the Models.tst template
//



export class UpdateTransactionResponseViewModel  { 
    success: boolean;
    transactionId: number;
    currencyId: number;
    grossTicketAmount: number;
    deliveryCharges: number;
    convenienceCharges: number;
    serviceCharge: number;
    discountAmount: number;
    netTicketAmount: number;
}